ISSD-385 rename
