+++ modules/dxp/apps/osb/osb-spring-boot/osb-spring-boot-zendesk-client/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay OSB Spring Boot Zendesk Client
+Bundle-SymbolicName: com.liferay.osb.spring.boot.zendesk.client
+Bundle-Version: 1.0.0
