+++ modules/dxp/apps/osb/osb-spring-boot/osb-spring-boot-zendesk-client/src/main/java/com/liferay/osb/spring/boot/zendesk/client/model/ZendeskUser.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.osb.client.extension.util.zendesk.model;
+package com.liferay.osb.spring.boot.zendesk.client.model;
 
 import org.json.JSONObject;
