+++ workspaces/liferay-customer-workspace/client-extensions/liferay-customer-etc-spring-boot/src/main/java/com/liferay/customer/TicketAttachmentsCompleteUploadRestController.java
@@ -7,8 +7,8 @@ package com.liferay.customer;
 
 import com.liferay.customer.object.model.TicketAttachment;
 import com.liferay.customer.object.service.TicketAttachmentWebService;
-import com.liferay.osb.client.extension.util.zendesk.model.ZendeskUser;
-import com.liferay.osb.client.extension.util.zendesk.service.ZendeskWebService;
+import com.liferay.osb.spring.boot.zendesk.client.model.ZendeskUser;
+import com.liferay.osb.spring.boot.zendesk.client.service.ZendeskWebService;
 import com.liferay.petra.string.CharPool;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.util.StringUtil;
